% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stan_data_brc.R
\name{add_age_strata_map}
\alias{add_age_strata_map}
\title{Map age to age strata}
\usage{
add_age_strata_map(stan_data, dt_contacts, single_contact_age = FALSE)
}
\arguments{
\item{stan_data}{A list object containing data for the Stan model.}

\item{dt_contacts}{A data table containing contacts information.}

\item{single_wave}{A logical value indicating whether the contact count data is from a
single wave survey (default: \code{TRUE}).}
}
\value{
The modified stan_data object with the age-to-strata map added.
}
\description{
This function maps age to age strata for a given survey and adds the resulting
map to the provided stan_data object. For COVIMOD survey, age is mapped to
predefined age strata. For POLYMOD survey, age is mapped to itself.
}
\examples{
data(stan_data)
stan_data <- add_age_strata_map(stan_data, survey = "COVIMOD")
stan_data <- add_age_strata_map(stan_data, survey = "POLYMOD")
}
