% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stan_data_brc.R
\name{add_contact_vector}
\alias{add_contact_vector}
\title{Add contact count vectors to the Stan data object}
\usage{
add_contact_vector(stan_data, dt_contacts)
}
\arguments{
\item{stan_data}{A list or data.frame object containing the parameters required
for the contact model.}

\item{dt_contacts}{A data.table object containing the contact count data to be used to
generate the contact count vectors.}
}
\value{
The Stan data object with contact count vectors added to it
}
\description{
This function adds contact count vectors to the Stan data object.
The contact vectors are generated from a data.table object of contact
data, which must contain columns for the age of the individuals, the age strata of
the contact, the gender of the individuals, the gender of the contact, and the
number of contacts between the individuals and the contacts. The function can
handle contact data from either the COVIMOD or POLYMOD surveys, and can add
contact vectors for single or multiple waves.
}
