% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stan_data_brc.R
\name{add_hsgp_parms}
\alias{add_hsgp_parms}
\title{Add HSGP parameters}
\usage{
add_hsgp_parms(stan_data, C, M1, M2)
}
\arguments{
\item{stan_data}{A list containing the Stan data.}

\item{C}{A scalar, the value of the HSGP hyperparameter C.}

\item{M1}{A scalar, the number of basis functions on the first dimension.}

\item{M2}{A scalar, the number of basis functions on the second dimension.}
}
\value{
The updated Stan data as a list.
}
\description{
Adds Hilbert Space approximate Gaussian Process (HSGP) parameters to the given Stan data.
}
\examples{
# Load sample data
data(stan_data)

# Add HSGP parameters
stan_data <- add_hsgp_parms(stan_data, C = 1.5, M1 = 30, M2 = 20)
}
