% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stan_data_brc.R
\name{add_N}
\alias{add_N}
\title{Add observation counts to the Stan data object}
\usage{
add_N(stan_data)
}
\description{
This function adds observation counts to the Stan data object. The observation
counts are generated from a data.table object of contact data, which must
}
