% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stan_data_brc.R
\name{add_non_nuisance_idx}
\alias{add_non_nuisance_idx}
\title{Add non-nuisance index}
\usage{
add_non_nuisance_idx(stan_data)
}
\arguments{
\item{stan_data}{A list containing the Stan data.}
}
\value{
A list containing the updated Stan data with the non-nuisance index added.
}
\description{
This function adds a non-nuisance index to the given Stan data.
}
\examples{
stan_data <- list(A = 3, B = 2, C = 4)
add_non_nuisance_idx(stan_data)
}
