% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stan_data_brc.R
\name{add_participant_offsets}
\alias{add_participant_offsets}
\title{Add Participant Offsets to Stan Data Object}
\usage{
add_participant_offsets(
  stan_data,
  dt_contacts,
  dt_offsets = NULL,
  single_wave = TRUE
)
}
\arguments{
\item{stan_data}{A Stan data object}

\item{dt_contacts}{A data table containing information about the contacts}

\item{dt_offsets}{A data table containing information about the participants (optional, defaults to NULL)}

\item{single_wave}{A logical value indicating whether the contact count data is from a
single wave survey (default: \code{TRUE}).}
}
\value{
A Stan data object with participant offsets added
}
\description{
This function adds participant offsets to a Stan data object based on the survey type and data tables containing information about the contacts and participants.
}
\examples{
add_part_offsets(stan_data, dt_contacts)
add_part_offsets(stan_data, dt_contacts, dt_offsets, survey = "POLYMOD")
}
