% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stan_data_brc.R
\name{add_population_offsets}
\alias{add_population_offsets}
\title{Add population offsets to Stan data}
\usage{
add_population_offsets(stan_data, dt_population, single_wave = TRUE)
}
\arguments{
\item{stan_data}{The data in the format expected by the Stan model}

\item{dt_population}{A data.table containing population sizes by gender and age}

\item{single_wave}{A logical value indicating whether the contact count data is from a
single wave survey (default: \code{TRUE}).}
}
\value{
The modified Stan data with population offsets added
}
\description{
Adds population offsets to Stan data for use in the Stan model. Population
offsets adjust for differences in population sizes between age groups and
genders.
}
