% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stan_data_brc.R
\name{add_row_major_idx}
\alias{add_row_major_idx}
\title{Add row-major index to Stan data object}
\usage{
add_row_major_idx(
  stan_data,
  dt_contacts,
  single_wave = TRUE,
  single_contact_age = FALSE
)
}
\arguments{
\item{stan_data}{Stan data object to be modified}

\item{dt_contacts}{Contact data in a data.table object}

\item{single_wave}{A logical value indicating whether the contact count data is from a
single wave survey (default: \code{TRUE}).}

\item{single_contact_age}{A logical value indicating whether the age of contact is in one year
intervals or aggregated into age groups (default: \code{FALSE})}
}
\value{
Stan data object with row-major index added
}
\description{
Adds a row-major index to the Stan data object based on the contact data in
dt_contacts. The index is stored in the Stan data object under the
appropriate column name depending on the survey type specified.
}
\details{
For "COVIMOD" surveyx, the row-major index is computed using the
formula (age_idx - 1) * 13 + age_strata_idx, where age_idx is the age of the
contact plus 1 and age_strata_idx is the age stratum of the contact. For
"POLYMOD" survey type, the row-major index is computed using the formula
(age_idx - 1) * 85 + age_strata_idx, where age_idx is the age of the contact
plus 1 and age_strata_idx is the age of the contact plus 1.
}
\examples{
data(stan_data)
data(dt_contacts)
stan_data <- add_row_major_idx(stan_data, dt_contacts, survey = "COVIMOD")
}
