% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stan_data_brc.R
\name{add_start_end_idx}
\alias{add_start_end_idx}
\title{Add start and end indices to a data frame based on contacts}
\usage{
add_start_end_idx(stan_data, dt_contacts)
}
\arguments{
\item{stan_data}{A data frame to which start and end indices are added.}

\item{dt_contacts}{A data table containing contacts information.}
}
\value{
A modified version of the stan_data with start and end indices added.
}
\description{
This function adds start and end indices for male and female genders in the given data
frame based on the contacts data. The function creates start and end indices for a given
gender using the \code{\link{make_start_end_idx}} function.
}
\examples{
# Load the data and add the start and end indices
data(contacts)
stan_data <- as.data.frame(contacts)
stan_data <- add_start_end_idx(stan_data, contacts)
}
