% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stan_data_brc.R
\name{add_std_age_idx}
\alias{add_std_age_idx}
\title{Add standardized age index}
\usage{
add_std_age_idx(stan_data)
}
\arguments{
\item{stan_data}{A list containing data needed for Stan modeling.
Must contain element \code{A}, which is the maximum age.}
}
\value{
A modified version of the input \code{stan_data} list, with
elements \code{age_idx_std} and \code{diff_idx_std} added.
}
\description{
Add standardized age index to the input \code{stan_data} list.
}
\examples{
add_std_age_idx(list(A = 100))
}
