% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stan_data_brc.R
\name{add_wave_repeat_u_map}
\alias{add_wave_repeat_u_map}
\title{Add wave-repeat to combined index map to Stan data}
\usage{
add_wave_repeat_u_map(stan_data)
}
\arguments{
\item{stan_data}{A list containing Stan data}
}
\value{
The modified Stan data with a new variable 'map_tr_to_u'
}
\description{
Given the number of waves (W) and the number of repeated participation (R),
create a map from the wave and repeat index to a combined index (U).
This optimization to improve Stan fitting time.
}
\examples{
stan_data <- list(U = 100, W = 4, R = 3)
stan_data <- add_wave_repeat_u_map(stan_data)
}
