% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_labels.R
\name{clean_labels}
\alias{clean_labels}
\title{Clean the labels of the variables}
\usage{
clean_labels(su, stan_data)
}
\arguments{
\item{su}{Posterior summary}

\item{stan_data}{The Stan data used to fit the model}
}
\description{
Clean the labels of the variables
}
