% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_labels_s2.R
\name{clean_labels_s2}
\alias{clean_labels_s2}
\title{Clean labels for the second stage of the variable selection analysis}
\usage{
clean_labels_s2(su, stan_data)
}
\arguments{
\item{su}{posterior summary}

\item{stan_data}{Stan data used to fit the model}
}
\value{
data.table
}
\description{
Clean labels for the second stage of the variable selection analysis
}
