% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cint_matrix.R
\name{extract_cint_matrix}
\alias{extract_cint_matrix}
\title{Extracts the posterior contact intensity matrix from the draws matrix}
\usage{
extract_cint_matrix(draws_matrix)
}
\arguments{
\item{draws_matrix}{}
}
\value{
A data.table with three columns: draws, variable, and value
}
\description{
Extracts the posterior contact intensity matrix from the draws matrix
}
