% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_missing_child_ages.R
\name{fill_missing_child_ages}
\alias{fill_missing_child_ages}
\title{Impute Missing Child Age Data}
\usage{
fill_missing_child_ages(dt, seed = 1527)
}
\arguments{
\item{dt}{A data.table object with columns for new_id, wave, age, and
age_strata}

\item{seed}{An optional integer for setting the random seed used in age
imputation}
}
\value{
A modified version of the input data.table object with missing ages
imputed
}
\description{
This function imputes missing age data for child records in a given data
table. The age is imputed using the age strata of the record and a uniform
distribution within that range.
}
\examples{
dt <- data.table(new_id = c(1, 2, 3, 4),
                              age = c(2, NA, NA, 5),
                              age_strata = c("0-4", "5-9", "5-9", "5-9"))
imputed_dt <- fill_missing_child_ages(dt)

}
