% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stan_data_brc.R
\name{init_stan_data}
\alias{init_stan_data}
\title{Create a template stan data object}
\usage{
init_stan_data(A = 85, C = 13)
}
\arguments{
\item{A}{Number of participant age groups}

\item{C}{Number of contact age groups}
}
\value{
list
}
\description{
Create a template stan data object
}
\examples{
stan_data <- init_stan_data()
}
