% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_age_gender_grid.R
\name{make_age_gender_grid}
\alias{make_age_gender_grid}
\title{Create a age gender grid}
\usage{
make_age_gender_grid(A)
}
\arguments{
\item{A}{The maximum age in the model}
}
\value{
A data table with demographic groups
}
\description{
This function generates a data.table with demographic groups that can be used as input
for a model. The data frame includes columns for age, age strata, and gender.
}
\examples{
make_grid(85)
}
