% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_age_strata_map_sim.R
\name{make_age_strata_map_sim}
\alias{make_age_strata_map_sim}
\title{Makes a matrix that maps continuous age to age strata for simulated data only}
\usage{
make_age_strata_map_sim(age_strata, A, C)
}
\arguments{
\item{age_strata}{A vector of age strata}

\item{A}{The number of age groups}

\item{C}{The number of age strata}
}
\value{
a matrix
}
\description{
Makes a matrix that maps continuous age to age strata for simulated data only
}
