% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_convergence_diagnostic_stats.R
\name{make_convergence_diagnostic_stats}
\alias{make_convergence_diagnostic_stats}
\title{A general function to make convergence diagnostic statistics}
\usage{
make_convergence_diagnostic_stats(fit, outdir = NA)
}
\arguments{
\item{fit}{A fitted CmdStanModel}

\item{outdir}{Directory to save outputs}
}
\value{
A summary table of estimates and diagnostics
}
\description{
A general function to make convergence diagnostic statistics
}
