% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_design_matrices.R
\name{make_design_matrices}
\alias{make_design_matrices}
\title{Make Design Matrix}
\usage{
make_design_matrices(df, analysis, model_type)
}
\arguments{
\item{df}{A data frame containing the variables to be processed.}
}
\value{
A design matrix with combined and processed columns.
}
\description{
This function creates a design matrix from a given data frame. It processes
specific columns of the data frame, handling categorical variables and
based on a repeat status vector.
}
\examples{
# Assuming df is a data frame with columns 'rep', 'age_strata', 'gender',
# 'hh_size', and 'job':
design_matrix <- make_design_matrix(df)

}
