% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_lowertri_idxset.R
\name{make_lowertri_idxset}
\alias{make_lowertri_idxset}
\title{Given a square index matrix, return the index set from the lower triangular part, including the diagonal}
\usage{
make_lowertri_idxset(A)
}
\arguments{
\item{A}{The size of one side of a squared matrix}
}
\value{
A matrix with two columns, where each row is a pair of indices
}
\description{
Given a square index matrix, return the index set from the lower triangular part, including the diagonal
}
\examples{
x <- make_lowertri_idxset(A)
X <- scale(x)
}
