% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_mh_data_varselect.R
\name{make_mh_data_varselect}
\alias{make_mh_data_varselect}
\title{Make stan data for variable selection models}
\usage{
make_mh_data_varselect(
  data,
  config,
  s1_results = NA,
  s2_results = NA,
  remove_first_dummy = TRUE
)
}
\arguments{
\item{data}{Preprocessed data}

\item{config}{Configuration list}
}
\value{
Stan data list
}
\description{
Make stan data for variable selection models
}
