% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nn_idxset.R
\name{make_nn_idxset}
\alias{make_nn_idxset}
\title{Compute the indices of the non-nuisance parameters for the difference in age parameterization}
\usage{
make_nn_idxset(A)
}
\arguments{
\item{A}{Number of age groups}
}
\value{
An index set. A matrix with two columns, where each row contains a pair of indices.
}
\description{
Compute the indices of the non-nuisance parameters for the difference in age parameterization
}
\examples{
nn_indices <- make_nn_idx_set(30)
X <- scale(nn_indices)
}
