% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_nn_lowertri_idxset.R
\name{make_nn_lowertri_idxset}
\alias{make_nn_lowertri_idxset}
\title{Make an index set representing the position of elements from a lower triangular matrix of non-nuisance parameters}
\usage{
make_nn_lowertri_idxset(A)
}
\arguments{
\item{A}{Age group}
}
\value{
A index set (matrix)
}
\description{
Make an index set representing the position of elements from a lower triangular matrix of non-nuisance parameters
}
\examples{
x <- make_nn_lowertri_idxset(85)
}
