% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_S_matrix.R
\name{make_S_matrix}
\alias{make_S_matrix}
\title{Make index matrix for eigenvalues and eigenfunctions}
\usage{
make_S_matrix(M1, M2)
}
\arguments{
\item{M1}{The number of basis functions for the first dimension}

\item{M2}{The number of basis functions for the second dimension}
}
\value{
A matrix with two columns, where each row is a pair of indices
}
\description{
Make index matrix for eigenvalues and eigenfunctions
}
\details{
For details, refer to section 3.2 of the paper "Practical Hilbert space approximate Bayesian Gaussian process for probabilistic programming"
by Riutort-Mayol et al. (2023).
}
