% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stan_data_brc.R
\name{make_stan_data}
\alias{make_stan_data}
\title{Make Stan data object for BRC model}
\usage{
make_stan_data(
  A,
  C,
  dt_contacts,
  dt_offsets,
  dt_population,
  model_params,
  waves = NULL,
  single_contact_age = FALSE
)
}
\arguments{
\item{A}{The number of participant age groups}

\item{C}{The number of contact age groups}

\item{dt_contacts}{A data.table object containing the contact count data}

\item{dt_offsets}{A data.table object containing the participant offset data}

\item{dt_population}{A data.table object containing the population data}

\item{model_params}{A list containing the model parameters}
}
\value{
Stan data object
}
\description{
Make Stan data object for BRC model
}
