% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stan_data_longit.R
\name{make_stan_data_longit}
\alias{make_stan_data_longit}
\title{Make Stan data for longitudinal data}
\usage{
make_stan_data_longit(data)
}
\arguments{
\item{data}{preprocessed data}
}
\value{
A Stan data list
}
\description{
Make Stan data for longitudinal data
}
