% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stan_data_varselect.R
\name{make_stan_data_varselect}
\alias{make_stan_data_varselect}
\title{Make stan data for variable selection models}
\usage{
make_stan_data_varselect(data, config)
}
\arguments{
\item{data}{Preprocessed data}

\item{config}{Configuration list}
}
\value{
Stan data list
}
\description{
Make stan data for variable selection models
}
