% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stan_data_varselect_s1.R
\name{make_stan_data_varselect_s1}
\alias{make_stan_data_varselect_s1}
\title{Create stan data object for the first stage of the variable selection analysis}
\usage{
make_stan_data_varselect_s1(
  data,
  u_vars = c("age_strata", "hh_size", "gender"),
  v_vars = c("job", "symp_none", "dow", "urbn_type")
)
}
\arguments{
\item{data}{data.table}
}
\value{
A list obejct to be passed to Stan
}
\description{
Create stan data object for the first stage of the variable selection analysis
}
