% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sym_from_lowertri_idx.R
\name{make_sym_from_lowertri_idx}
\alias{make_sym_from_lowertri_idx}
\title{Indices to make a symmetric matrix from a lower triangular matrix}
\usage{
make_sym_from_lowertri_idx(A)
}
\arguments{
\item{A}{}
}
\value{
A vector of indices
}
\description{
Indices to make a symmetric matrix from a lower triangular matrix
}
\details{
Suppose there is a vector containing the elements of a lower triangular matrix arranged in column-major order.
This function generates the indices required to obtain a vector containing the elements of a symmetric full matrix,
generated by copying the lower triangular matrix to the upper triangular part.
}
