% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_baseline_effects.R
\name{plot_baseline_effects}
\alias{plot_baseline_effects}
\title{Plot baseline effects of the variable selection model}
\usage{
plot_baseline_effects(fit, stan_data, config, outdir = NA)
}
\arguments{
\item{fit}{CmdStanR fit object}

\item{stan_data}{Stan data used to fit the model}

\item{config}{Configuration list}

\item{outdir}{Output directory}
}
\value{
A ggplot object
}
\description{
Plot baseline effects of the variable selection model
}
