% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc_age_job.R
\name{preproc_age_job}
\alias{preproc_age_job}
\title{Preprocess the age_strata and job columns to avoid complete colinearity}
\usage{
preproc_age_job(data, drop.na = TRUE)
}
\arguments{
\item{data}{Participant data}
}
\value{
A data frame or date table with the age_strata and job columns preprocessed
}
\description{
Preprocess the age_strata and job columns to avoid complete colinearity
}
