% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc_gam_data.R
\name{preproc_gam_data}
\alias{preproc_gam_data}
\title{Preprocess the COVIMOD data for the GAM model}
\usage{
preproc_gam_data(w, part, hh, nhh, nuts)
}
\arguments{
\item{w}{wave}

\item{part}{participant data (data.table)}

\item{hh}{household contacts (data.table)}

\item{nhh}{non-household contacts (data.table)}

\item{nuts}{NUTS regions (data.table)}
}
\value{
processed contact data (data.table)
}
\description{
Preprocess the COVIMOD data for the GAM model
}
