% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify_age.R
\name{stratify_age}
\alias{stratify_age}
\title{Stratify age into predefined age strata}
\usage{
stratify_age(dt, name = "age", strat_name = "age_strata")
}
\arguments{
\item{dt}{A data.table object.}

\item{name}{The name of the column containing age values.}

\item{strat_name}{The desired name for the stratified column.}
}
\value{
A data.table object with the original age column stratified into
age strata according to the predefined ranges.
}
\description{
This function takes a data.table object and stratifies the \code{name} column
into predefined age strata using the \code{cut} function from the \code{stats}
package. The strata are defined as follows:
}
\details{
\describe{
\item{\code{0-4}}{Ages 0 to 4}
\item{\code{5-9}}{Ages 5 to 9}
\item{\code{10-14}}{Ages 10 to 14}
\item{\code{15-19}}{Ages 15 to 19}
\item{\code{20-24}}{Ages 20 to 24}
\item{\code{25-34}}{Ages 25 to 34}
\item{\code{35-44}}{Ages 35 to 44}
\item{\code{45-54}}{Ages 45 to 54}
\item{\code{55-64}}{Ages 55 to 64}
\item{\code{65-69}}{Ages 65 to 69}
\item{\code{70-74}}{Ages 70 to 74}
\item{\code{75-79}}{Ages 75 to 79}
\item{\code{80-84}}{Ages 80 to 84}
\item{\code{85+}}{Ages 85 and over}
}
}
\examples{
dt <- data.table(age = c(23, 56, 70, 81))
stratify_age(dt)

}
