% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_cint_marginal.R
\name{summarise_cint_marginal}
\alias{summarise_cint_marginal}
\title{Summarise posterior distribution of marginal contact intensity}
\usage{
summarise_cint_marginal(
  cint_matrix,
  use_quantile = TRUE,
  probs = c(0.5, 0.025, 0.25, 0.75, 0.975),
  labels = c("M", "CL", "Q25", "Q75", "CU")
)
}
\arguments{
\item{cint_matrix}{A data.table containing the posterior distribution of contact intensity}

\item{use_quantile}{A logical indicating whether to summarise using quantiles}

\item{probs}{A vector of quantiles to summarise}

\item{labels}{A vector of labels for the quantiles}
}
\value{
A data.table containing the summary statistics posterior distribution of marginal contact intensity
}
\description{
Summarise posterior distribution of marginal contact intensity
}
