% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_cint_matrix.R
\name{summarise_cint_matrix}
\alias{summarise_cint_matrix}
\title{Summarise the posterior distribution of the contact intensity matrix}
\usage{
summarise_cint_matrix(
  cint_matrix,
  probs = c(0.5, 0.025, 0.25, 0.75, 0.975),
  labels = c("M", "CL", "Q25", "Q50", "CU")
)
}
\arguments{
\item{cint_matrix}{A data.table containing the posterior draws of the contact intensity matrix}

\item{probs}{A vector of quantiles to compute}

\item{labels}{A vector of labels for the quantiles}
}
\value{
A data.table containing the posterior quantiles of the contact intensity matrix
}
\description{
Summarise the posterior distribution of the contact intensity matrix
}
