% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_wcint.R
\name{summarise_wcint}
\alias{summarise_wcint}
\title{Summarise the weighted contact intensity of different population subgroups}
\usage{
summarise_wcint(fit)
}
\arguments{
\item{fit}{A fit object from cmdstanr}
}
\value{
A data.table with the median, 2.5th and 97.5th percentiles of the contact intensity
}
\description{
Summarise the weighted contact intensity of different population subgroups
}
