% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_wcint.R
\name{summarise_wcint_65p}
\alias{summarise_wcint_65p}
\title{Summarise Weighted Contact Intensity for 65+ Age Group}
\usage{
summarise_wcint_65p(draws_log_mu, draws_beta_sex, draws_beta_hhsize)
}
\arguments{
\item{draws_log_mu}{Numeric matrix of posterior draws of the log-transformed baseline contact intensities.}

\item{draws_beta_sex}{Numeric matrix of posterior draws of coefficients for the effect of sex on contact intensity.}

\item{draws_beta_hhsize}{Numeric matrix of posterior draws of coefficients for the effect of household size on contact intensity.}
}
\value{
A data.table containing summary measures of the exponentiated contact intensities,
including the median and the 2.5th and 97.5th percentiles, with an additional column
"variable" set to "65+".
}
\description{
This function calculates the weighted contact intensity for individuals aged 65 and above.
It uses the posterior draws of the log-transformed baseline contact intensities adjusted for
age, sex, and household size. The resulting draws are exponentiated and summarised by computing
the median and the 2.5th and 97.5th percentiles.
}
\examples{
\dontrun{
  result <- summarise_wcint_65p(draws_log_mu, draws_beta_sex, draws_beta_hhsize)
  print(result)
}
}
