% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_wcint.R
\name{summarise_wcint_all}
\alias{summarise_wcint_all}
\title{Calculate Population Average Contact Intensity}
\usage{
summarise_wcint_all(draws_log_mu, draws_beta_sex, draws_beta_hhsize)
}
\arguments{
\item{draws_log_mu}{Numeric matrix of posterior draws of the log-transformed baseline contact intensities.}

\item{draws_beta_sex}{Numeric matrix of posterior draws of coefficients for the effect of sex on contact intensity.}

\item{draws_beta_hhsize}{Numeric matrix of posterior draws of coefficients for the effect of household size on contact intensity.}
}
\value{
A data.table containing the summarized draws of the exponential of adjusted contact intensity.
The table includes summary measures and quantiles (2.5\% and 97.5\%) for the contact intensity,
with an additional column "variable" set to "all".
}
\description{
This function calculates the population average contact intensity weighted by age, sex, and household size.
}
\examples{
\dontrun{
  result <- summarise_wcint_all(draws_log_mu, draws_beta_sex, draws_beta_hhsize)
  print(result)
}
}
