% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_wcint.R
\name{summarise_wcint_female}
\alias{summarise_wcint_female}
\title{Summarise Weighted Contact Intensity for Females}
\usage{
summarise_wcint_female(draws_log_mu, draws_beta_sex)
}
\arguments{
\item{draws_log_mu}{Numeric matrix of posterior draws of the log-transformed baseline contact intensities.}

\item{draws_beta_sex}{Numeric matrix of posterior draws for the effect of sex on contact intensity.}
}
\value{
A data.table containing summary measures of the exponentiated contact intensities for females,
including the median, 2.5th and 97.5th percentiles, with an additional column "variable" set to "female".
}
\description{
This function calculates the weighted contact intensity for females by adjusting the baseline
log contact intensities for the female effect and incorporating age and gender-based population weights.
}
