% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_wcint.R
\name{summarise_wcint_self_employed}
\alias{summarise_wcint_self_employed}
\title{Summarise Weighted Contact Intensity for Self-Employed Individuals}
\usage{
summarise_wcint_self_employed(
  draws_log_mu,
  draws_beta_sex,
  draws_beta_hhsize,
  draws_beta_job,
  stan_data
)
}
\arguments{
\item{draws_log_mu}{Numeric matrix of posterior draws of the log-transformed baseline contact intensities.}

\item{draws_beta_sex}{Numeric matrix of posterior draws of coefficients for the effect of sex on contact intensity.}

\item{draws_beta_hhsize}{Numeric matrix of posterior draws of coefficients for the effect of household size on contact intensity.}

\item{draws_beta_job}{Numeric matrix of posterior draws of coefficients for the effect of job type on contact intensity.}

\item{stan_data}{A list containing Stan data, including "X_job", the design matrix for job effects.}
}
\value{
A data.table containing summary measures of the exponentiated contact intensities for self-employed individuals,
including the median, and 2.5th and 97.5th percentiles, with an additional column "variable" set to "self-employed".
}
\description{
This function calculates the weighted contact intensity for self-employed individuals by adjusting
the baseline contact intensity for individual effects of sex and household size, and adding the job effect
corresponding to self-employment. The resulting contact intensity is exponentiated and summarised using
median and selected quantiles.
}
\examples{
\dontrun{
  result <- summarise_wcint_self_employed(draws_log_mu, draws_beta_sex, draws_beta_hhsize,
                                          draws_beta_job, stan_data)
  print(result)
}
}
