% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_wcint.R
\name{summarise_wcint_toddler}
\alias{summarise_wcint_toddler}
\title{Summarise Weighted Contact Intensity for Toddlers}
\usage{
summarise_wcint_toddler(draws_log_mu, draws_beta_sex, draws_beta_hhsize)
}
\arguments{
\item{draws_log_mu}{Numeric matrix of posterior draws of the log-transformed baseline contact intensities.}

\item{draws_beta_sex}{Numeric matrix of posterior draws of coefficients for the effect of sex on contact intensity.}

\item{draws_beta_hhsize}{Numeric matrix of posterior draws of coefficients for the effect of household size on contact intensity.}
}
\value{
A data.table containing summary measures of the exponentiated contact intensities,
including the median, 2.5th and 97.5th percentiles, with an additional column "variable" set to "toddlers".
}
\description{
This function summarizes the weighted contact intensity for toddlers by calculating
the exponential of the adjusted contact intensities. It uses the posterior draws of the
log-transformed baseline contact intensities and adjusts for sex and household size.
}
\examples{
\dontrun{
  result <- summarise_wcint_toddler(draws_log_mu, draws_beta_sex, draws_beta_hhsize)
  print(result)
}
}
