% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_wcint.R
\name{summarise_wcint_unemployed}
\alias{summarise_wcint_unemployed}
\title{Summarise Weighted Contact Intensity for Unemployed Individuals}
\usage{
summarise_wcint_unemployed(
  draws_log_mu,
  draws_beta_sex,
  draws_beta_hhsize,
  draws_beta_job,
  stan_data
)
}
\arguments{
\item{draws_log_mu}{Numeric matrix of posterior draws of the log-transformed baseline contact intensities.}

\item{draws_beta_sex}{Numeric matrix of posterior draws for the effect of sex on contact intensity.}

\item{draws_beta_hhsize}{Numeric matrix of posterior draws for the effect of household size on contact intensity.}

\item{draws_beta_job}{Numeric matrix of posterior draws for the effect of job type on contact intensity.}

\item{stan_data}{A list containing Stan data, including "X_job", the design matrix for job effects.}
}
\value{
A data.table containing the summary measures of the exponentiated contact intensity, including the median,
2.5th and 97.5th percentiles, with an additional column "variable" set to "unemployed".
}
\description{
This function calculates the weighted contact intensity for unemployed individuals by computing the exponential
of the adjusted contact intensities. It uses the posterior draws of the log-transformed baseline contact intensity,
and adjusts for sex and household size. In addition, it incorporates the job effects for unemployed individuals,
averaging over those unemployed who are looking versus not looking for work.
}
