% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_stan_data.R
\name{update_stan_data}
\alias{update_stan_data}
\title{Update the stan data for the current step of the stepwise variable selection process}
\usage{
update_stan_data(i, stage, stan_data, post_init)
}
\arguments{
\item{i}{The current index of the variable to include}

\item{stage}{The variable selection analysis stage}

\item{stan_data}{The stan data list}

\item{post_init}{The ranked posterior summary of the initial screening model}
}
\value{
An update stan data object
}
\description{
Update the stan data for the current step of the stepwise variable selection process
}
