% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_stan_data_varselect.R
\name{update_stan_data_varselect}
\alias{update_stan_data_varselect}
\title{Update stan_data object with selected variables}
\usage{
update_stan_data_varselect(stan_data, stan_fit)
}
\arguments{
\item{stan_data}{the initial stan_data object (list)}

\item{stan_fit}{the stan_fit object}
}
\value{
A new stan_data object with the selected variables
}
\description{
Update stan_data object with selected variables
}
