% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varselect_s2_initial.R
\name{varselect_s2_initial}
\alias{varselect_s2_initial}
\title{Run initial screening for stage one of the variable selection analysis}
\usage{
varselect_s2_initial(data, config, s1_results)
}
\arguments{
\item{data}{A data table containing the preprocessed contact data}

\item{config}{A list containing the configurations for the variable selection analysis}

\item{s1_results}{A list containing the results from stage one of the variable selection analysis}
}
\value{
A data table of the posterior summary of the regression coefficients sorted by their effect size
}
\description{
Run initial screening for stage one of the variable selection analysis
}
