% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varselect_s2_stepwise.R
\name{varselect_s2_stepwise}
\alias{varselect_s2_stepwise}
\title{Run stepwise variable selection}
\usage{
varselect_s2_stepwise(data, config, s1_results)
}
\arguments{
\item{data}{A data frame with the contact data}

\item{config}{A list with configuration parameters}
}
\value{
A list with the posterior summary of the best fitting model and the LOO-ELPD of each model
}
\description{
Run stepwise variable selection
}
