% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcint.R
\name{wcint_age_gender}
\alias{wcint_age_gender}
\title{Weighted average of contact intensity by age and gender}
\usage{
wcint_age_gender(age_set, draws_log_m, draws_beta)
}
\arguments{
\item{age_set}{A vector of age groups}

\item{draws_log_m}{A draws_matrix of log contact intensity}

\item{draws_beta}{A draws_matrix of beta (fixed effects)}
}
\value{
A vector of draws for the weighted contact intensity
}
\description{
Weighted average of contact intensity by age and gender
}
