% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcint.R
\name{wcint_agh}
\alias{wcint_agh}
\title{Weighted average of contact intensity by age, gender, and household size}
\usage{
wcint_agh(age_set, draws_log_m, draws_beta)
}
\arguments{
\item{age_set}{A vector of age groups}

\item{draws_log_m}{A draws_matrix of log contact intensity}

\item{draws_beta}{A draws_matrix of beta (fixed effects)}
}
\description{
Weighted average of contact intensity by age, gender, and household size
}
