% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcint.R
\name{wcint_aghu}
\alias{wcint_aghu}
\title{Weighted average of contact intensity by age, gender, household size, urban-rural typography}
\usage{
wcint_aghu(
  age_set,
  draws_log_mu,
  draws_beta_sex,
  draws_beta_hhsize,
  draws_beta_urbn
)
}
\arguments{
\item{age_set}{A vector of age groups}

\item{draws_log_mu}{A draws_matrix of log contact intensity}

\item{draws_beta_sex}{A draws_matrix of beta_sex (fixed effects)}

\item{draws_beta_hhsize}{A draws_matrix of beta_hhsize (fixed effects)}

\item{draws_beta_urbn}{A draws_matrix of beta_urbn (fixed effects)}
}
\value{
A draws_array for the weighted contact intensity
}
\description{
Weighted average of contact intensity by age, gender, household size, urban-rural typography
}
