% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcint.R
\name{wcint_hh_size}
\alias{wcint_hh_size}
\title{Weighted average of contact intensity by household size}
\usage{
wcint_hh_size(draws_log_m, draws_beta)
}
\arguments{
\item{draws_log_m}{a draws_matrix of log contact intensity, usually the output of wcint_age_gender.}

\item{draws_beta}{a draws_matrix of beta (fixed effects)}
}
\value{
A vector of draws for the weighted contact intensity
}
\description{
Weighted average of contact intensity by household size
}
