% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcint.R
\name{wcint_hhsize}
\alias{wcint_hhsize}
\title{Weighted average of contact intensity by household size}
\usage{
wcint_hhsize(log_wcint, draws_beta_hhsize)
}
\arguments{
\item{log_wcint}{a vector of log weighted contact intensity, usually the output of wcint_age_gender.}

\item{draws_beta}{a draws_matrix of beta (fixed effects)}
}
\value{
A vector of draws for the weighted contact intensity
}
\description{
Weighted average of contact intensity by household size
}
