% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_labels.R
\name{clean_labels}
\alias{clean_labels}
\title{Clean the labels of the variables}
\usage{
clean_labels(x)
}
\arguments{
\item{x}{A character vector of variable names}
}
\value{
A list of variables and labels
}
\description{
Clean the labels of the variables
}
