% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dummy_matrix.R
\name{make_dummy_matrix}
\alias{make_dummy_matrix}
\title{Make dummy variable matrix}
\usage{
make_dummy_matrix(data, variable, include = NULL, ...)
}
\arguments{
\item{data}{Participant data}

\item{variable}{Variable to make dummy variables for}

\item{include}{Columns to include in the output matrix}

\item{...}{Additional arguments to pass to fastDummies::dummy_cols}
}
\value{
A matrix of dummy variables for the specified variable
}
\description{
Make dummy variable matrix
}
