% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stan_data_varselect_s2.R
\name{make_stan_data_varselect_s2}
\alias{make_stan_data_varselect_s2}
\title{Make Stan data object for second stage of the variable selection}
\usage{
make_stan_data_varselect_s2(data, selected_vars, fit)
}
\arguments{
\item{data}{data.table}

\item{selected_vars}{character vector of variables selected in the first stage}

\item{fit}{cmdstanr fit object}
}
\value{
Stan data object (list)
}
\description{
Make Stan data object for second stage of the variable selection
}
